requirements.txt
pandas
matplotlib
streamlit
seaborn
dotenv
xlwings
google-generativeai